feat: Separated API for business logic
