+++ backend/assembler/src/main/java/module-info.java
@@ -3,9 +3,14 @@
  */
 module assembler_module{
 
-    requires persistence_module;
-    requires businesslogic_module;
     requires datarecords_module;
+    
+    requires persistence_api_module;
+    requires persistence_module;
+    
+    requires businesslogic_api_module;
+    requires businesslogic_module;
+     
     requires restapi_module;
     
     requires java.logging;
