+++ backend/businesslogic-api/pom.xml
@@ -6,22 +6,16 @@
         <artifactId>airlineinformationsystem</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>businesslogic</artifactId>
+    <artifactId>businesslogic-api</artifactId>
+    <packaging>jar</packaging>
     <dependencies>
         <dependency>
-            <groupId>io.github.fontysvenlo.ais</groupId>
+            <groupId>${project.groupId}</groupId>
             <artifactId>datarecords</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>persistence-api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>businesslogic-api</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
+    <properties>
+        <exec.mainClass>io.github.fontysvenlo.ais.businesslogic.api.BusinesslogicApi</exec.mainClass>
+    </properties>
 </project>
