+++ backend/businesslogic-api/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * Module businesslogic_api.
+ */
+module businesslogic_api_module {
+    
+    requires datarecords_module;
+    
+    exports io.github.fontysvenlo.ais.businesslogic.api;
+}
