+++ backend/businesslogic/src/main/java/io/github/fontysvenlo/ais/businesslogic/BusinessLogicAPIImpl.java
@@ -1,5 +1,7 @@
 package io.github.fontysvenlo.ais.businesslogic;
 
+import io.github.fontysvenlo.ais.businesslogic.api.BusinessLogicAPI;
+import io.github.fontysvenlo.ais.businesslogic.api.CustomerManager;
 import io.github.fontysvenlo.ais.persistence.api.PersistenceAPI;
 
 /**
@@ -23,6 +25,6 @@ class BusinessLogicAPIImpl implements BusinessLogicAPI {
      */
     @Override
     public CustomerManager getCustomerManager() {
-        return new CustomerManager(persistenceAPI.getCustomerRepository());
+        return new CustomerManagerImpl(persistenceAPI.getCustomerRepository());
     }
 }
