+++ backend/businesslogic/src/main/java/module-info.java
@@ -4,6 +4,7 @@
 module businesslogic_module {   
     requires datarecords_module;
     requires persistence_api_module;
+    requires businesslogic_api_module;
     
     exports io.github.fontysvenlo.ais.businesslogic;
 }
