+++ backend/restapi/pom.xml
@@ -17,11 +17,6 @@
             <version>6.4.0</version>
         </dependency>
         <dependency>
-            <groupId>io.github.fontysvenlo.ais</groupId>
-            <artifactId>businesslogic</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <version>2.0.16</version>
@@ -35,6 +30,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-annotations</artifactId>
             <version>2.18.0</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>businesslogic-api</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
